+++ nemo_text_processing/text_normalization/en/taggers/date.py
@@ -13,17 +13,13 @@
 # limitations under the License.
 
 import pynini
-from nemo_text_processing.text_normalization.en.utils import (
-    augment_labels_with_punct_at_end,
-    get_abs_path,
-    load_labels,
-)
+from nemo_text_processing.text_normalization.en.utils import TO_LOWER, get_abs_path
+from nemo_text_processing.tn_itn_utils.file_utils import augment_labels_with_punct_at_end, load_labels
 from nemo_text_processing.tn_itn_utils.graph_utils import (
     NEMO_CHAR,
     NEMO_DIGIT,
     NEMO_LOWER,
     NEMO_SIGMA,
-    TO_LOWER,
     GraphFst,
     delete_extra_space,
     delete_space,
