+++ nemo_text_processing/text_normalization/en/taggers/decimal.py
@@ -13,7 +13,8 @@
 # limitations under the License.
 
 import pynini
-from nemo_text_processing.tn_itn_utils.graph_utils import NEMO_SIGMA, TO_UPPER, GraphFst, get_abs_path
+from nemo_text_processing.text_normalization.en.utils import TO_UPPER, get_abs_path
+from nemo_text_processing.tn_itn_utils.graph_utils import NEMO_SIGMA, GraphFst
 from pynini.lib import pynutil
 
 delete_space = pynutil.delete(" ")
