+++ nemo_text_processing/text_normalization/en/taggers/electronic.py
@@ -14,14 +14,8 @@
 
 
 import pynini
-from nemo_text_processing.tn_itn_utils.graph_utils import (
-    NEMO_ALPHA,
-    NEMO_DIGIT,
-    NEMO_SIGMA,
-    GraphFst,
-    get_abs_path,
-    insert_space,
-)
+from nemo_text_processing.text_normalization.en.utils import get_abs_path
+from nemo_text_processing.tn_itn_utils.graph_utils import NEMO_ALPHA, NEMO_DIGIT, NEMO_SIGMA, GraphFst, insert_space
 from pynini.lib import pynutil
