+++ nemo_text_processing/text_normalization/en/taggers/fraction.py
@@ -13,7 +13,8 @@
 # limitations under the License.
 
 import pynini
-from nemo_text_processing.tn_itn_utils.graph_utils import GraphFst, get_abs_path
+from nemo_text_processing.text_normalization.en.utils import get_abs_path
+from nemo_text_processing.tn_itn_utils.graph_utils import GraphFst
 from pynini.lib import pynutil
