+++ nemo_text_processing/text_normalization/en/taggers/punctuation.py
@@ -16,7 +16,8 @@ import sys
 from unicodedata import category
 
 import pynini
-from nemo_text_processing.text_normalization.en.utils import get_abs_path, load_labels
+from nemo_text_processing.text_normalization.en.utils import get_abs_path
+from nemo_text_processing.tn_itn_utils.file_utils import load_labels
 from nemo_text_processing.tn_itn_utils.graph_utils import NEMO_NOT_SPACE, NEMO_SIGMA, GraphFst
 from pynini.examples import plurals
 from pynini.lib import pynutil
