+++ nemo_text_processing/text_normalization/en/taggers/telephone.py
@@ -13,7 +13,7 @@
 # limitations under the License.
 
 import pynini
-from nemo_text_processing.text_normalization.en.utils import get_abs_path
+from nemo_text_processing.text_normalization.en.utils import get_abs_path, plurals
 from nemo_text_processing.tn_itn_utils.graph_utils import (
     NEMO_ALPHA,
     NEMO_DIGIT,
@@ -22,7 +22,6 @@ from nemo_text_processing.tn_itn_utils.graph_utils import (
     delete_extra_space,
     delete_space,
     insert_space,
-    plurals,
 )
 from pynini.lib import pynutil
