+++ nemo_text_processing/tn_itn_utils/file_utils.py
@@ -1,4 +1,4 @@
-# Copyright (c) 2021, NVIDIA CORPORATION.  All rights reserved.
+# Copyright (c) 2021, NVIDIA CORPORATION  All rights reserved.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -13,19 +13,9 @@
 # limitations under the License.
 
 import csv
-import os
+from typing import Union
 
-
-def get_abs_path(rel_path):
-    """
-    Get absolute path
-
-    Args:
-        rel_path: relative path to this file
-        
-    Returns absolute path
-    """
-    return os.path.dirname(os.path.abspath(__file__)) + '/' + rel_path
+import inflect
 
 
 def load_labels(abs_path):
@@ -58,3 +48,21 @@ def augment_labels_with_punct_at_end(labels):
             if label[0][-1] == "." and label[1][-1] != ".":
                 res.append([label[0], label[1] + "."] + label[2:])
     return res
+
+
+_inflect = inflect.engine()
+
+
+def num_to_word(x: Union[str, int]):
+    """
+    converts integer to spoken representation
+
+    Args
+        x: integer
+
+    Returns: spoken representation
+    """
+    if isinstance(x, int):
+        x = str(x)
+        x = _inflect.number_to_words(str(x)).replace("-", " ").replace(",", "")
+    return x
